third_party/WebKit/LayoutTests/crypto/aes-cbc-parseAlgorithm-failures.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests bad algorithm inputs for AES-CBC");

jsTestIsAsync = true;

var keyData = hexStringToUint8Array("2b7e151628aed2a6abf7158809cf4f3c");
var data = asciiToUint8Array("hello");
var key = null;

Promise.resolve(null).then(function(result) {
    var usages = ['encrypt', 'decrypt'];
    var extractable = false;
    var algorithm = {name: 'aes-cbc'};

    return crypto.subtle.importKey('raw', keyData, algorithm, extractable, usages);
}).then(function(result) {
    key = result;

    // Bad iv
    return crypto.subtle.encrypt({name: 'AES-CBC', iv: null}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    // Missing iv
    return crypto.subtle.decrypt({name: 'AES-CBC'}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    // iv is not a buffer
    return crypto.subtle.encrypt({name: 'AES-CBC', iv: 3}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    // iv is too short
    return crypto.subtle.encrypt({name: 'AES-CBC', iv: new Uint8Array(0)}, key, data);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
